Fix potential uncaught strtol errors in commandline parsing on 64bit machines
Found by gcc warning about always true comparisions.
(cherry picked from commit 7561d8b9753e5992b611bd1321e18132bd35bcaa)

+++ undertaker/undertaker.cpp
@@ -364,7 +364,7 @@ int main (int argc, char ** argv) {
     char *worklist = NULL;
     char *whitelist = NULL;
 
-    int threads = 1;
+    long threads = 1;
     std::list<std::string> models;
     std::string main_model = "x86";
     /* Default is dead/undead analysis */
